Replace static/singleton RoutineUIComposer with a subclassable instance initialization
This was necessary in order to keep the dependencies required to makeCreateRoutine not necessary to be passed as an argument for navigation. This would have required that RoutineRepository be Hashable, which is not ideal, as it is creating a Repository change for UI requirements. (Also its not easy to do).

Moving it from the static methods allowed us to hold a reference to only routine repository dependency that we would need to inject to the views that need it.

However a singleton was not ideal to test on, so it was necessary to open the class and use instance implementation in order to use a subclass in the unit tests where views were instantiated with the routine repository spy